refactor(daemon-service): own shutdown coordination
