+++ crates/tracedecay/src/mcp/server/connection.rs
@@ -6,208 +6,6 @@ use super::*;
 pub(super) const MAX_CONCURRENT_CONNECTION_READS: usize =
     crate::daemon::MAX_CONCURRENT_REQUESTS_PER_DAEMON_CLIENT;
 
-pub(in crate::mcp::server) struct McpShutdownCompletion {
-    state: Arc<McpShutdownState>,
-}
-
-#[derive(Default)]
-struct McpShutdownState {
-    running: AtomicBool,
-    terminal: std::sync::Mutex<Option<crate::daemon::ShutdownStatus>>,
-    coordinator_task: tokio::sync::Mutex<Option<tokio::task::JoinHandle<()>>>,
-    changed: tokio::sync::Notify,
-}
-
-struct McpShutdownCoordinatorCompletion(Arc<McpShutdownState>);
-
-impl Drop for McpShutdownCoordinatorCompletion {
-    fn drop(&mut self) {
-        self.0.changed.notify_waiters();
-    }
-}
-
-impl Default for McpShutdownCompletion {
-    fn default() -> Self {
-        Self {
-            state: Arc::new(McpShutdownState::default()),
-        }
-    }
-}
-
-impl McpShutdownCompletion {
-    #[hotpath::skip]
-    async fn coordinate_until<Work>(
-        &self,
-        deadline: tokio::time::Instant,
-        work: Work,
-    ) -> crate::daemon::ShutdownStatus
-    where
-        Work: std::future::Future<Output = crate::daemon::ShutdownStatus> + Send + 'static,
-    {
-        let mut work = Some(work);
-        loop {
-            self.join_finished_coordinator().await;
-            if !self.state.running.load(Ordering::Acquire) {
-                self.wait_for_finished_coordinator().await;
-                self.join_finished_coordinator().await;
-            }
-            if let Some(status) = self.terminal_status() {
-                self.wait_for_finished_coordinator().await;
-                self.join_finished_coordinator().await;
-                return status;
-            }
-
-            let mut coordinator_task = self.state.coordinator_task.lock().await;
-            if coordinator_task.is_some() {
-                let running = self.state.running.load(Ordering::Acquire);
-                drop(coordinator_task);
-                if running {
-                    return self.wait_for_terminal_status_until(deadline).await;
-                }
-                self.wait_for_finished_coordinator().await;
-                self.join_finished_coordinator().await;
-                continue;
-            }
-            if self
-                .state
-                .running
-                .compare_exchange(false, true, Ordering::AcqRel, Ordering::Acquire)
-                .is_err()
-            {
-                drop(coordinator_task);
-                return self.wait_for_terminal_status_until(deadline).await;
-            }
-            if let Some(status) = self.terminal_status() {
-                self.state.running.store(false, Ordering::Release);
-                drop(coordinator_task);
-                self.wait_for_finished_coordinator().await;
-                self.join_finished_coordinator().await;
-                return status;
-            }
-
-            let Some(work) = work.take() else {
-                self.state.finish(crate::daemon::ShutdownStatus::Failed(
-                    "MCP shutdown coordinator lost its work future".to_owned(),
-                ));
-                drop(coordinator_task);
-                return crate::daemon::ShutdownStatus::Failed(
-                    "MCP shutdown coordinator lost its work future".to_owned(),
-                );
-            };
-            let state = Arc::clone(&self.state);
-            let task = tokio::spawn(async move {
-                let _completion = McpShutdownCoordinatorCompletion(Arc::clone(&state));
-                let runner = tokio::spawn(work);
-                let status = match runner.await {
-                    Ok(status) => status,
-                    Err(error) => crate::daemon::ShutdownStatus::Failed(error.to_string()),
-                };
-                state.finish(status);
-            });
-            *coordinator_task = Some(task);
-            drop(coordinator_task);
-            return self.wait_for_terminal_status_until(deadline).await;
-        }
-    }
-
-    #[hotpath::skip]
-    async fn join_finished_coordinator(&self) {
-        let result = {
-            let mut coordinator_task = self.state.coordinator_task.lock().await;
-            let Some(task) = coordinator_task.as_mut() else {
-                return;
-            };
-            if !task.is_finished() {
-                return;
-            }
-            let result = task.await;
-            coordinator_task.take();
-            result
-        };
-        if let Err(error) = result {
-            tracing::error!(%error, "MCP shutdown coordinator task failed after receipt");
-            self.state
-                .finish(crate::daemon::ShutdownStatus::Failed(error.to_string()));
-        }
-    }
-
-    #[hotpath::skip]
-    async fn wait_for_finished_coordinator(&self) {
-        loop {
-            let notified = self.state.changed.notified();
-            tokio::pin!(notified);
-            notified.as_mut().enable();
-            let finished = self
-                .state
-                .coordinator_task
-                .lock()
-                .await
-                .as_ref()
-                .is_none_or(tokio::task::JoinHandle::is_finished);
-            if finished {
-                return;
-            }
-            notified.as_mut().await;
-        }
-    }
-
-    fn terminal_status(&self) -> Option<crate::daemon::ShutdownStatus> {
-        self.state
-            .terminal
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner)
-            .clone()
-    }
-
-    #[hotpath::skip]
-    async fn wait_for_terminal_status_until(
-        &self,
-        deadline: tokio::time::Instant,
-    ) -> crate::daemon::ShutdownStatus {
-        loop {
-            if let Some(status) = self.terminal_status() {
-                self.wait_for_finished_coordinator().await;
-                self.join_finished_coordinator().await;
-                return status;
-            }
-            if !self.state.running.load(Ordering::Acquire) {
-                self.wait_for_finished_coordinator().await;
-                self.join_finished_coordinator().await;
-                return crate::daemon::ShutdownStatus::TimedOut;
-            }
-            let notified = self.state.changed.notified();
-            tokio::pin!(notified);
-            notified.as_mut().enable();
-            if let Some(status) = self.terminal_status() {
-                self.wait_for_finished_coordinator().await;
-                self.join_finished_coordinator().await;
-                return status;
-            }
-            if !self.state.running.load(Ordering::Acquire) {
-                self.wait_for_finished_coordinator().await;
-                self.join_finished_coordinator().await;
-                return crate::daemon::ShutdownStatus::TimedOut;
-            }
-            if tokio::time::timeout_at(deadline, notified).await.is_err() {
-                return crate::daemon::ShutdownStatus::TimedOut;
-            }
-        }
-    }
-}
-
-impl McpShutdownState {
-    fn finish(&self, status: crate::daemon::ShutdownStatus) {
-        if status != crate::daemon::ShutdownStatus::TimedOut {
-            *self
-                .terminal
-                .lock()
-                .unwrap_or_else(std::sync::PoisonError::into_inner) = Some(status);
-        }
-        self.running.store(false, Ordering::Release);
-        self.changed.notify_waiters();
-    }
-}
-
 /// The single production context the portable MCP connection scheduler uses.
 ///
 /// It borrows the already-constructed server authorities; it does not own a
@@ -1516,128 +1314,5 @@ mod cancellable_queue_tests {
             .expect("join selected cancellation connection")
             .expect("serve selected cancellation connection");
         harness.shutdown().await;
-    }
-}
-
-#[cfg(test)]
-mod shutdown_tests {
-    use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-
-    use super::*;
-
-    struct RetainedShutdownOwner(Arc<AtomicBool>);
-
-    impl Drop for RetainedShutdownOwner {
-        fn drop(&mut self) {
-            self.0.store(true, Ordering::Release);
-        }
-    }
-
-    #[tokio::test]
-    async fn cancelled_shutdown_waiter_does_not_cancel_owned_work() {
-        let completion = Arc::new(McpShutdownCompletion::default());
-        let attempts = Arc::new(AtomicUsize::new(0));
-        let entered = Arc::new(tokio::sync::Notify::new());
-        let (release, released) = tokio::sync::oneshot::channel();
-
-        let first_completion = Arc::clone(&completion);
-        let first_attempts = Arc::clone(&attempts);
-        let first_entered = Arc::clone(&entered);
-        let first = tokio::spawn(async move {
-            first_completion
-                .coordinate_until(
-                    tokio::time::Instant::now() + Duration::from_secs(5),
-                    async move {
-                        first_attempts.fetch_add(1, Ordering::AcqRel);
-                        first_entered.notify_one();
-                        let _ = released.await;
-                        crate::daemon::ShutdownStatus::Clean
-                    },
-                )
-                .await
-        });
-        entered.notified().await;
-        first.abort();
-        assert!(
-            first
-                .await
-                .expect_err("cancel first shutdown waiter")
-                .is_cancelled()
-        );
-
-        release.send(()).expect("release retained shutdown work");
-        let retry_attempts = Arc::clone(&attempts);
-        let retry = completion
-            .coordinate_until(
-                tokio::time::Instant::now() + Duration::from_secs(1),
-                async move {
-                    retry_attempts.fetch_add(1, Ordering::AcqRel);
-                    panic!("retry must await the retained shutdown work");
-                },
-            )
-            .await;
-
-        assert_eq!(retry, crate::daemon::ShutdownStatus::Clean);
-        assert_eq!(attempts.load(Ordering::Acquire), 1);
-    }
-
-    #[tokio::test(start_paused = true)]
-    async fn timed_out_shutdown_retains_work_until_retry_observes_terminal_status() {
-        let completion = Arc::new(McpShutdownCompletion::default());
-        let attempts = Arc::new(AtomicUsize::new(0));
-        let owner_dropped = Arc::new(AtomicBool::new(false));
-        let entered = Arc::new(tokio::sync::Notify::new());
-        let (release, released) = tokio::sync::oneshot::channel();
-
-        let first_completion = Arc::clone(&completion);
-        let first_attempts = Arc::clone(&attempts);
-        let first_entered = Arc::clone(&entered);
-        let first_owner_dropped = Arc::clone(&owner_dropped);
-        let first = tokio::spawn(async move {
-            first_completion
-                .coordinate_until(
-                    tokio::time::Instant::now() + Duration::from_secs(1),
-                    async move {
-                        let _owner = RetainedShutdownOwner(first_owner_dropped);
-                        first_attempts.fetch_add(1, Ordering::AcqRel);
-                        first_entered.notify_one();
-                        let _ = released.await;
-                        crate::daemon::ShutdownStatus::Clean
-                    },
-                )
-                .await
-        });
-        entered.notified().await;
-        tokio::time::advance(Duration::from_secs(1)).await;
-        assert_eq!(
-            first.await.expect("first timed-out shutdown"),
-            crate::daemon::ShutdownStatus::TimedOut
-        );
-        assert!(
-            !owner_dropped.load(Ordering::Acquire),
-            "the timed-out attempt must retain its owner for a retry"
-        );
-
-        let retry_completion = Arc::clone(&completion);
-        let retry_attempts = Arc::clone(&attempts);
-        let retry = tokio::spawn(async move {
-            retry_completion
-                .coordinate_until(
-                    tokio::time::Instant::now() + Duration::from_secs(1),
-                    async move {
-                        retry_attempts.fetch_add(1, Ordering::AcqRel);
-                        panic!("retry must await the retained shutdown owner");
-                    },
-                )
-                .await
-        });
-        release.send(()).expect("release retained shutdown owner");
-
-        assert_eq!(
-            retry.await.expect("retry shutdown"),
-            crate::daemon::ShutdownStatus::Clean
-        );
-        assert_eq!(attempts.load(Ordering::Acquire), 1);
-        assert!(owner_dropped.load(Ordering::Acquire));
     }
 }
